requirements.txt
python-dotenv
openai
discord.py